# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Traffic-light verdict.
 */
export type DocumentCheckRecommendation =
  (typeof DocumentCheckRecommendation)[keyof typeof DocumentCheckRecommendation];

export const DocumentCheckRecommendation = {
  ready_to_notarize: "ready_to_notarize",
  fix_first: "fix_first",
  needs_review: "needs_review",
} as const;

/**
 * The Docsy service the inspector recommends booking.
 */
export type DocumentCheckSuggestedService =
  (typeof DocumentCheckSuggestedService)[keyof typeof DocumentCheckSuggestedService];

export const DocumentCheckSuggestedService = {
  ron: "ron",
  mobile: "mobile",
  "in-office": "in-office",
  apostille: "apostille",
  "loan-signing": "loan-signing",
} as const;

export type DocumentCheckDateField =
  (typeof DocumentCheckDateField)[keyof typeof DocumentCheckDateField];

export const DocumentCheckDateField = {
  present: "present",
  missing: "missing",
  unknown: "unknown",
} as const;

export interface DocumentCheckNotarialBlock {
  present: boolean;
  /**
   * Human-readable position, e.g. "page 2, bottom".
   * @nullable
   */
  location: string | null;
}

export interface DocumentCheckResult {
  documentType: string;
  notarialBlockPresent: DocumentCheckNotarialBlock;
  signaturePresent: boolean;
  dateField: DocumentCheckDateField;
  redFlags: string[];
  recommendation: DocumentCheckRecommendation;
  recommendationRationale: string;
  suggestedService: DocumentCheckSuggestedService;
  /** Data URL of the original image for the result preview. Empty string for PDFs. */
  thumbnailDataUrl: string;
  /**
   * App Storage object path, or null if storage was unavailable.
   * @nullable
   */
  storedObjectPath: string | null;
  /**
   * ISO timestamp at which the stored object becomes eligible for
deletion. Cleanup is best-effort (sweeps run hourly and after
each new upload) and may run a few hours after this time.

   * @nullable
   */
  storedExpiresAt: string | null;
  /**
   * UUID identifying this scan in App Storage. Pass to
/document-check/save-to-vault to keep the scan permanently.
Null when the upload could not be persisted.

   * @nullable
   */
  scanId?: string | null;
}

export interface DocumentCheckResponse {
  ok: boolean;
  result: DocumentCheckResult;
}

export interface DocumentCheckError {
  ok: boolean;
  error: string;
}

export type VaultFileServiceType =
  (typeof VaultFileServiceType)[keyof typeof VaultFileServiceType];

export const VaultFileServiceType = {
  ron: "ron",
  mobile: "mobile",
  loan: "loan",
  apostille: "apostille",
  court: "court",
} as const;

export interface VaultFile {
  id: string;
  name: string;
  serviceType: VaultFileServiceType;
  serviceLabel: string;
  /** YYYY-MM-DD */
  date: string;
  /** Human-readable, e.g. '247 KB' */
  size: string;
}

export interface SaveScanToVaultResponse {
  ok: boolean;
  file: VaultFile;
}

export type DocumentCheckBody = {
  /** The document to inspect (PDF, JPG, or PNG, ≤25 MB). */
  file: Blob;
};

export type SaveScanToVaultBody = {
  /** The scanId returned by a previous /document-check response. */
  scanId: string;
};