# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: document-check
    description: Pre-flight document inspection (free, anonymous)
  - name: vault
    description: Signed-in member document vault (Safe+)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /document-check:
    post:
      operationId: documentCheck
      tags: [document-check]
      summary: Run AI pre-flight inspection on an uploaded document
      description: |
        Accepts a PDF, JPG, or PNG of a document the user intends to notarize
        and returns structured findings: document type, presence of notarial
        block / signature line / date field, red flags, and a one-of-five
        suggested service. Anonymous uploads are stored in App Storage and
        scheduled for deletion ~24 hours later (best-effort, enforced by a
        periodic and request-driven cleanup sweep).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The document to inspect (PDF, JPG, or PNG, ≤25 MB).
      responses:
        "200":
          description: Inspection complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
        "503":
          description: Inspection service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
  /document-check/save-to-vault:
    post:
      operationId: saveScanToVault
      tags: [document-check, vault]
      summary: Promote an anonymous scan into the signed-in user's Safe+ vault
      description: |
        Requires a Bearer session token. Takes the `scanId` returned by a
        recent /document-check call and copies the stored original +
        result.json out of the 24-hour anonymous-storage prefix into the
        user's permanent vault prefix, then appends a VaultFile entry.
        Idempotent: re-saving the same scanId replaces the existing
        vault entry rather than duplicating it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scanId]
              properties:
                scanId:
                  type: string
                  description: The scanId returned by a previous /document-check response.
      responses:
        "200":
          description: Saved to vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveScanToVaultResponse"
        "400":
          description: Invalid scanId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
        "404":
          description: Scan no longer available (TTL expired)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
        "500":
          description: Save failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCheckError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DocumentCheckRecommendation:
      type: string
      enum: [ready_to_notarize, fix_first, needs_review]
      description: Traffic-light verdict.
    DocumentCheckSuggestedService:
      type: string
      enum: [ron, mobile, in-office, apostille, loan-signing]
      description: The Docsy service the inspector recommends booking.
    DocumentCheckDateField:
      type: string
      enum: [present, missing, unknown]
    DocumentCheckNotarialBlock:
      type: object
      required: [present, location]
      properties:
        present:
          type: boolean
        location:
          type: [string, "null"]
          description: Human-readable position, e.g. "page 2, bottom".
    DocumentCheckResult:
      type: object
      required:
        - documentType
        - notarialBlockPresent
        - signaturePresent
        - dateField
        - redFlags
        - recommendation
        - recommendationRationale
        - suggestedService
        - thumbnailDataUrl
        - storedObjectPath
        - storedExpiresAt
      properties:
        documentType:
          type: string
        notarialBlockPresent:
          $ref: "#/components/schemas/DocumentCheckNotarialBlock"
        signaturePresent:
          type: boolean
        dateField:
          $ref: "#/components/schemas/DocumentCheckDateField"
        redFlags:
          type: array
          items:
            type: string
        recommendation:
          $ref: "#/components/schemas/DocumentCheckRecommendation"
        recommendationRationale:
          type: string
        suggestedService:
          $ref: "#/components/schemas/DocumentCheckSuggestedService"
        thumbnailDataUrl:
          type: string
          description: Data URL of the original image for the result preview. Empty string for PDFs.
        storedObjectPath:
          type: [string, "null"]
          description: App Storage object path, or null if storage was unavailable.
        storedExpiresAt:
          type: [string, "null"]
          format: date-time
          description: |
            ISO timestamp at which the stored object becomes eligible for
            deletion. Cleanup is best-effort (sweeps run hourly and after
            each new upload) and may run a few hours after this time.
        scanId:
          type: [string, "null"]
          description: |
            UUID identifying this scan in App Storage. Pass to
            /document-check/save-to-vault to keep the scan permanently.
            Null when the upload could not be persisted.
    DocumentCheckResponse:
      type: object
      required: [ok, result]
      properties:
        ok:
          type: boolean
          enum: [true]
        result:
          $ref: "#/components/schemas/DocumentCheckResult"
    DocumentCheckError:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
    VaultFile:
      type: object
      required: [id, name, serviceType, serviceLabel, date, size]
      properties:
        id:           { type: string }
        name:         { type: string }
        serviceType:
          type: string
          enum: [ron, mobile, loan, apostille, court]
        serviceLabel: { type: string }
        date:         { type: string, description: "YYYY-MM-DD" }
        size:         { type: string, description: "Human-readable, e.g. '247 KB'" }
    SaveScanToVaultResponse:
      type: object
      required: [ok, file]
      properties:
        ok:
          type: boolean
          enum: [true]
        file:
          $ref: "#/components/schemas/VaultFile"