# lib/api-zod/src/generated/types/documentCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentCheckDateField } from "./documentCheckDateField";
import type { DocumentCheckNotarialBlock } from "./documentCheckNotarialBlock";
import type { DocumentCheckRecommendation } from "./documentCheckRecommendation";
import type { DocumentCheckSuggestedService } from "./documentCheckSuggestedService";

export interface DocumentCheckResult {
  documentType: string;
  notarialBlockPresent: DocumentCheckNotarialBlock;
  signaturePresent: boolean;
  dateField: DocumentCheckDateField;
  redFlags: string[];
  recommendation: DocumentCheckRecommendation;
  recommendationRationale: string;
  suggestedService: DocumentCheckSuggestedService;
  /** Data URL of the original image for the result preview. Empty string for PDFs. */
  thumbnailDataUrl: string;
  /**
   * App Storage object path, or null if storage was unavailable.
   * @nullable
   */
  storedObjectPath: string | null;
  /**
   * ISO timestamp at which the stored object becomes eligible for
deletion. Cleanup is best-effort (sweeps run hourly and after
each new upload) and may run a few hours after this time.

   * @nullable
   */
  storedExpiresAt: Date | null;
  /**
   * UUID identifying this scan in App Storage. Pass to
/document-check/save-to-vault to keep the scan permanently.
Null when the upload could not be persisted.

   * @nullable
   */
  scanId?: string | null;
}