# lib/api-zod/src/generated/types/saveScanToVaultResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VaultFile } from "./vaultFile";

export interface SaveScanToVaultResponse {
  ok: boolean;
  file: VaultFile;
}